requirements.txt
fastapi
uvicorn[standard]
transformers
torch
omegaconf
python-multipart
python-docx
pypdf
requests
